+++ paper_admin/static/paper_admin/src/components/back-button/back-button.js
@@ -1,7 +1,13 @@
-document.addEventListener("click", function (event) {
+document.addEventListener("click", event => {
     const link = event.target.closest(".back-button");
     if (link) {
         event.preventDefault();
-        window.history.back();
+
+        const params = new URLSearchParams(window.location.search);
+        if (params.has("_popup")) {
+            window.close(); // Close the popup.
+        } else {
+            window.history.back(); // Otherwise, go back.
+        }
     }
 });