+++ paper_admin/static/paper_admin/src/components/related-object-lookups/related-object-lookups.js
@@ -3,61 +3,29 @@
 // raw_id_fields.
 // -------------------------------------------
 
-function showAdminPopup(triggeringLink, name_regexp, add_popup) {
-    const name = triggeringLink.id.replace(name_regexp, "");
-    let href = triggeringLink.href;
-    if (add_popup) {
-        if (href.indexOf("?") === -1) {
-            href += "?_popup=1";
-        } else {
-            href += "&_popup=1";
-        }
-    }
+import popupUtils from "js/utilities/popup_utils.js";
 
-    // Центрирование popup
-    const browser_left = typeof window.screenX !== "undefined" ? window.screenX : window.screenLeft,
-        browser_top = typeof window.screenY !== "undefined" ? window.screenY : window.screenTop,
-        browser_width = typeof window.outerWidth !== "undefined" ? window.outerWidth : document.body.clientWidth,
-        browser_height = typeof window.outerHeight !== "undefined" ? window.outerHeight : document.body.clientHeight,
-        popup_width = Math.max(940, browser_width * 0.7),
-        popup_height = Math.max(600, browser_height * 0.7),
-        top_position = browser_top + Math.round((browser_height - popup_height) / 2),
-        left_position = browser_left + Math.round((browser_width - popup_width) / 2);
-
-    const win = window.open(
-        href,
-        name,
-        "width=" +
-            popup_width +
-            ",height=" +
-            popup_height +
-            ",top=" +
-            top_position +
-            ",left=" +
-            left_position +
-            ",resizable=yes,scrollbars=yes"
-    );
-
-    win.focus();
-    return false;
-}
 
 function showRelatedObjectLookupPopup(triggeringLink) {
-    return showAdminPopup(triggeringLink, /^lookup_/, true);
+    return popupUtils.showAdminPopup(triggeringLink, /^lookup_/, true);
 }
 
 function dismissRelatedLookupPopup(win, chosenId) {
-    const name = win.name;
+    const name = popupUtils.removePopupIndex(win.name);
     const elem = document.getElementById(name);
+
     if (elem.classList.contains("vManyToManyRawIdAdminField") && elem.value) {
         elem.value += "," + chosenId;
     } else {
-        document.getElementById(name).value = chosenId;
+        elem.value = chosenId;
     }
+
+    popupUtils.removeRelatedWindow(win);
     win.close();
 }
 
-document.addEventListener("click", function (event) {
+document.addEventListener("click", event => {
+    // Клик на кнопку поиска объекта
     const triggeringLink = event.target.closest(".related-lookup");
     if (triggeringLink) {
         event.preventDefault();
@@ -68,6 +36,7 @@ document.addEventListener("click", function (event) {
         }
     }
 
+    // Клик на объект из списка
     const dissmissLink = event.target.closest("a[data-popup-opener]");
     if (dissmissLink) {
         event.preventDefault();
@@ -77,5 +46,3 @@ document.addEventListener("click", function (event) {
 
 window.showRelatedObjectLookupPopup = showRelatedObjectLookupPopup;
 window.dismissRelatedLookupPopup = dismissRelatedLookupPopup;
-
-export { showAdminPopup };