+++ paper_admin/static/paper_admin/src/js/changeform.js
@@ -1,3 +1,4 @@
+/* global gettext */
 import formUtils from "js/utilities/form_utils.js";
 import { Select2Widget } from "components/select2";
 import { InlineFormset } from "bem/paper-formset/paper-formset.js";
@@ -13,23 +14,36 @@ select2_changeform.initAll(".select-field select");
 
 // Инициализация inline-форм
 let formsets = [];
-document.querySelectorAll(".paper-formset").forEach(function (element) {
+document.querySelectorAll(".paper-formset").forEach(element => {
     const formset = new InlineFormset(element);
     formset.updateButtonsState();
     formsets.push(formset);
 });
 
+// Предотвращение повторного сохранения
+let submitted = false;
+const form = document.querySelector(".paper-form");
+form.addEventListener("submit", event => {
+    event.preventDefault();
+    if (submitted) {
+        const answer = window.confirm(gettext("You have already submitted this form. Are you sure you want to submit it again?"));
+        if (!answer) {
+            return;
+        }
+    }
+    event.target.submit();
+    submitted = true;
+});
+
 // Установка значения поля сортировки перед сохранением.
 // Назначить сортировку сразу нельзя из-за того, что extra-формы не должны меняться.
-document.addEventListener("submit", function () {
-    formsets.forEach(function (formset) {
+document.addEventListener("submit", () => {
+    formsets.forEach(formset => {
         let index = 0;
-        formset.getForms().forEach(
-            function (form) {
-                if (form.classList.contains("has_original") || formUtils.containsChangedField(form)) {
-                    this.setFormOrder(form, index++);
-                }
-            }.bind(formset)
-        );
+        formset.getForms().forEach(form => {
+            if (form.classList.contains("has_original") || formUtils.containsChangedField(form)) {
+                formset.setFormOrder(form, index++);
+            }
+        });
     });
 });