+++ paper_admin/static/paper_admin/src/widgets/text-field/text-field.js
@@ -12,7 +12,7 @@ class AutosizeWidget extends Widget {
 }
 
 // FIX bootstrap tabs
-$(document).on("shown.bs.tab", function (event) {
+$(document).on("shown.bs.tab", event => {
     const tab_selector = event.target.getAttribute("href");
     const tab_pane = tab_selector && document.querySelector(tab_selector);
     if (tab_pane) {